chore: refactored payments into the steps folder
