+++ src/main/modules/oidc/index.ts
@@ -3,8 +3,17 @@ import { Application, NextFunction, Response } from 'express';
 
 import { CALLBACK_URL, getRedirectUrl, getUserDetails } from '../../app/auth/user/oidc';
 import { getCaseApi } from '../../app/case/CaseApi';
+import { PaymentStatus, State } from '../../app/case/definition';
 import { AppRequest } from '../../app/controller/AppRequest';
-import { SIGN_IN_URL, SIGN_OUT_URL } from '../../steps/urls';
+import { PaymentModel } from '../../app/payment/PaymentModel';
+import {
+  APPLICATION_SUBMITTED,
+  PAYMENT_CALLBACK_URL,
+  PAYMENT_URL,
+  PageLink,
+  SIGN_IN_URL,
+  SIGN_OUT_URL,
+} from '../../steps/urls';
 
 /**
  * Adds the oidc middleware to add oauth authentication
@@ -41,6 +50,38 @@ export class OidcMiddleware {
         } else {
           res.redirect(SIGN_IN_URL);
         }
+      })
+    );
+
+    app.use(
+      errorHandler(async (req: AppRequest, res: Response, next: NextFunction) => {
+        if ([PAYMENT_URL, PAYMENT_CALLBACK_URL].includes(req.path as PageLink)) {
+          return next();
+        }
+
+        if (
+          [State.Submitted, State.AwaitingDocuments, State.AwaitingHWFDecision].includes(req.session.userCase.state) &&
+          req.path !== APPLICATION_SUBMITTED
+        ) {
+          return res.redirect(APPLICATION_SUBMITTED);
+        }
+
+        const payments = new PaymentModel(req.session.userCase.payments);
+        if (payments.hasPayment) {
+          const lastPaymentAttempt = payments.lastPayment;
+          if (
+            req.session.userCase.state === State.AwaitingPayment &&
+            lastPaymentAttempt.paymentStatus === PaymentStatus.IN_PROGRESS
+          ) {
+            return res.redirect(PAYMENT_CALLBACK_URL);
+          }
+
+          if (lastPaymentAttempt.paymentStatus === PaymentStatus.SUCCESS && req.path !== APPLICATION_SUBMITTED) {
+            return res.redirect(APPLICATION_SUBMITTED);
+          }
+        }
+
+        return next();
       })
     );
   }
